Changed wayback.urlprefix to be relative rather than absolute
This is to fix https://github.com/nla/govau/issues/25 -
Original website page 'warning' message popping up

The pm.gov.au site detects any links on the page which are absolute
and assumes they are links to external sites. It marks all such
links and then produces a Javascript 'confirm' dialog every time
the user clicks on one of them. Unfortunately, because Wayback
was rewriting all URLs to be absolute the pm.gov.au site was riddled
with false positives.