Fix for nla/govau issues #28 and #30.
Added UrlRewriteFilter JAR to lib/
Added mapping for UrlRewriteFilter of /* (i.e. all requests)
Rewrite rules in WEB-INF/urlrewrite.xml catch invalid URLs
i.e. those not starting with /wayback/, /WEB-INF/, /favicon.ico
or /js and rewrites them to /invalidUrl/<original URL>.

The companion WAR file invalidUrl.war should be dropped into the
same folder as the ROOT deployment of Wayback. This will then
be deployed to a context of /invalidUrl and simply sends a
'Resource not in archive' message with a 404 Not Found status
as the response.